requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
APScheduler==3.10.4
reportlab==4.2.2
Werkzeug==3.0.3
itsdangerous==2.2.0
click==8.1.7
gunicorn==22.0.0
requests==2.32.3
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
